Made password grant_type application requirement optional to allow project self-authentication
